Change a bunch. Potentially break "reddit@" with non-default header. Tests will have to be run.
Changed blinky animation to be stepped, taking less CPU power (or so I've been told). Further changed it to be a unicode block character whose color changes. It now also happens to follow DOS's cursor blink speed. Changed after text to be a single text size, instead of an easy-to-break scaling font size that I don't have a clue how it even worked in the first place. In my tests with injecting the CSS directly to the /r/eliteshell theme sub, it seems to have broken the "reddit@" text because of a non-standard header image. This could be a simple fix of changing it back to default by removing the custom (easter egg?) image. I feel like I'm forgetting something from the attempt at cherry-picking fixes and not putting in DOS-specific changes from the variation I'm working on, but I'll make a seperate branch for the sake of showing the changes.